requirements.txt
flask==2.3.3
line-bot-sdk==3.6.0
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0
gevent==23.9.1 